refactor: 추상화 계층 분리
<before>
- Model
- View
- Controller

<after>
- Model
  - RacingCarGameModel
- View
  - RacingCarGameView
- Controller
  - RacingCarGameController
