+++ src/js/core/Controller.js
@@ -1,0 +1,6 @@
+export default class Controller {
+  constructor(model, view) {
+    this.model = model;
+    this.view = view;
+  }
+}
