+++ src/js/core/Model.js
@@ -1,0 +1,13 @@
+import { cloneObject } from '../utils/utils.js';
+
+export default class Model {
+  state;
+
+  setState(newState) {
+    this.state = { ...this.state, ...newState };
+  }
+
+  generatePayload() {
+    return cloneObject(this.state);
+  }
+}
