+++ src/js/core/View.js
@@ -1,0 +1,20 @@
+import { $all } from '../utils/utils.js';
+
+export default class View {
+  constructor(template) {
+    this.template = template;
+  }
+
+  bindEventListener(type, selector, callback) {
+    const children = [...$all(selector)];
+    const isTarget = (target) =>
+      children.includes(target) || target.closest(selector);
+
+    this.$target.addEventListener(type, (e) => {
+      if (!isTarget(e.target)) return;
+
+      e.preventDefault();
+      callback(e);
+    });
+  }
+}
